requirements/server.txt
faster-whisper==1.1.0
websockets
onnxruntime==1.17.0
numba
kaldialign
soundfile
ffmpeg-python
scipy
jiwer
evaluate
numpy<2
openai-whisper==20240930
tokenizers==0.20.3
python-dotenv
scikit-learn
pyannote-audio

librosa
openvino
openvino-genai
openvino-tokenizers
optimum 
optimum-intel